CI: move some functionality from lesson.component to another files
