+++ ci/src/app/_components/topic/lesson/index.ts
@@ -1,1 +1,5 @@
 ﻿export * from './lesson.component';
+export * from './good-dialog/good-dialog.component';
+export * from './bad-dialog/bad-dialog.component';
+export * from './report-dialog/report-dialog.component';
+export * from './question-with-chart/question-with-chart.component';
