+++ ci/src/app/_components/topic/lesson/lesson.component.css
@@ -16,12 +16,3 @@
   margin: 4px;
   color: #fff;
 }
-
-#chart-container{
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-direction: column;
-  margin: 0;
-  padding: 0;
-}
